agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bedrock-dq-agent"
version = "0.2.0"
description = "AI Data Quality Agent v2 - Agentic DQ Monitoring with profiling, impact analysis, and root cause tracing"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=1.27.0",
    "strands-agents-tools>=0.1.0",
    "boto3>=1.35.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "slack-sdk>=3.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "moto[dynamodb,s3]>=5.0",
    "respx>=0.21",
    "ruff>=0.8",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ai_dq_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 120
